Create FunctionQueue
Function queue made to avoid warnings appearing when updating data stores. Essential with the DataService as it uses SetAsync a lot more.

Issue: In theory, there is a chance a player can leave and re-join while the request to save the data is still in the queue. A solution to this could be a priority system which would go as follows:

1: Saving Data on Leaving
2: Updating Global Leaderboards (quite demanding on DataStore budget stuff)
3: Autosaving